+++ limber-backend-application/module/forms/api/src/main/kotlin/io/limberapp/backend/module/forms/service/formTemplate/FormTemplateQuestionService.kt
@@ -10,7 +10,7 @@ interface FormTemplateQuestionService {
         formTemplatePartId: UUID,
         formTemplateQuestionGroupId: UUID,
         model: FormTemplateQuestionModel,
-        index: Short? = null
+        index: Int? = null
     )
 
     fun update(
