+++ limber-backend-application/module/forms/app/src/main/kotlin/io/limberapp/backend/module/forms/service/formTemplate/FormTemplateQuestionServiceImpl.kt
@@ -16,7 +16,7 @@ internal class FormTemplateQuestionServiceImpl @Inject constructor(
         formTemplatePartId: UUID,
         formTemplateQuestionGroupId: UUID,
         model: FormTemplateQuestionModel,
-        index: Short?
+        index: Int?
     ) {
         val entity = formTemplateQuestionMapper.entity(model)
         formTemplateQuestionStore.create(formTemplateId, formTemplatePartId, formTemplateQuestionGroupId, entity)
